# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ClinicAI SaaS multi-tenant WhatsApp automation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clinics:
    get:
      operationId: listClinics
      tags: [clinics]
      summary: List all clinics
      responses:
        "200":
          description: List of clinics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Clinic"
    post:
      operationId: createClinic
      tags: [clinics]
      summary: Create a clinic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClinicBody"
      responses:
        "201":
          description: Created clinic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clinic"
  /clinics/{id}:
    get:
      operationId: getClinic
      tags: [clinics]
      summary: Get a clinic by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Clinic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clinic"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateClinic
      tags: [clinics]
      summary: Update clinic AI settings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClinicBody"
      responses:
        "200":
          description: Updated clinic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clinic"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Clinic:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        apiKey:
          type: string
        aiName:
          type: string
        aiPersonalityPrompt:
          type: string
        knowledgeBase:
          type: string
        clinicType:
          type: string
          enum: [medical, vet, dental, beauty, education, retail, food, technology, services, other]
        aiExternalApiKey:
          type: string
          nullable: true
          description: "Chave de API de IA externa (retornada mascarada: ****XXXX)"
        asaasApiKey:
          type: string
          nullable: true
          description: "Chave de integração Asaas (retornada mascarada: ****XXXX)"
        whatsappPhoneNumberId:
          type: string
          nullable: true
          description: "ID do número WhatsApp Business API"
        whatsappAccessToken:
          type: string
          nullable: true
          description: "Token de acesso WhatsApp Business API (retornado mascarado: ****XXXX)"
        evolutionInstanceName:
          type: string
          nullable: true
        whatsappProvider:
          type: string
          enum: [evolution, meta]
          description: "Provedor WhatsApp ativo: evolution (padrão) ou meta"
        mercadoPagoAccessToken:
          type: string
          nullable: true
          description: "Access Token Mercado Pago (retornado mascarado: ****XXXX)"
        schedulingEnabled:
          type: boolean
          description: "Se false, o módulo de agendamento está desativado"
        aiEnabled:
          type: boolean
          description: "Se false, a IA não responde automaticamente pelo WhatsApp"
        isBlocked:
          type: boolean
          description: "Se true, a conta está bloqueada pelo superadmin"
        blockedReason:
          type: string
          nullable: true
          description: "Motivo do bloqueio"
        blockedAt:
          type: string
          format: date-time
          nullable: true
          description: "Data em que a conta foi bloqueada"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, phone, apiKey, aiName, aiPersonalityPrompt, knowledgeBase, clinicType, aiEnabled, whatsappProvider, schedulingEnabled, isBlocked, createdAt, updatedAt]
    CreateClinicBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        clinicType:
          type: string
          enum: [medical, vet, dental, beauty, education, retail, food, technology, services, other]
        aiName:
          type: string
        aiPersonalityPrompt:
          type: string
        knowledgeBase:
          type: string
      required: [name, phone, clinicType, aiName, aiPersonalityPrompt, knowledgeBase]
    UpdateClinicBody:
      type: object
      properties:
        name:
          type: string
        aiName:
          type: string
        aiPersonalityPrompt:
          type: string
        knowledgeBase:
          type: string
        clinicType:
          type: string
          enum: [medical, vet, dental, beauty, education, retail, food, technology, services, other]
        aiExternalApiKey:
          type: string
          nullable: true
          description: "Chave de API de IA externa — enviar o valor completo para atualizar"
        asaasApiKey:
          type: string
          nullable: true
          description: "Chave de integração Asaas — enviar o valor completo para atualizar"
        whatsappPhoneNumberId:
          type: string
          nullable: true
        whatsappAccessToken:
          type: string
          nullable: true
          description: "Token de acesso WhatsApp Business API — enviar o valor completo para atualizar"
        aiEnabled:
          type: boolean
          description: "Liga/desliga o atendimento automático por IA"
        whatsappProvider:
          type: string
          enum: [evolution, meta]
          description: "Provedor WhatsApp ativo"
        evolutionInstanceName:
          type: string
          nullable: true
        mercadoPagoAccessToken:
          type: string
          nullable: true
          description: "Access Token Mercado Pago — enviar o valor completo para atualizar"
        schedulingEnabled:
          type: boolean
          description: "Liga/desliga o módulo de agendamento"
...